fix and sweaten gossips

+++ data/sql/db-characters/zone_difficulty_instance_saves.sql
@@ -8,7 +8,7 @@ CREATE TABLE IF NOT EXISTS `zone_difficulty_instance_saves` (
 CREATE TABLE IF NOT EXISTS zone_difficulty_hardmode_score(
     `GUID` INT NOT NULL DEFAULT 0,
     `Type` TINYINT NOT NULL DEFAULT 0,
-    `Score` TINYINT NOT NULL DEFAULT 0,
+    `Score` INT NOT NULL DEFAULT 0,
     PRIMARY KEY (`GUID`, `Type`)
 );
